recruitx-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
python-dotenv
pandas
openpyxl
pdfplumber
python-multipart
